+++ R/get_tiles.R
@@ -8,7 +8,54 @@ guess_format <- function(x) {
   c("png", "jpg")[grepl("satellite", x) + 1L]
 }
 
-#' Download Mapbox imagery tiles
+
+#' Unpack Mapbox terrain-RGB
+#'
+#' Mapbox terrain-rgb stores global elevation packed into Byte integers. 
+#' 
+#' This function unpacks the three layers of a raster to give floating point elevation data. 
+#' @param x three layer raster object
+#' @param filename optional, filename to store the output
+#'
+#' @return terra rast object with one numeric layer
+#' @export
+#'
+#' @examples
+#' if (interactive() && !is.null(get_api_key())) {
+#' unpack_rgb(read_tiles(type = "mapbox.terrain-rgb"))
+#' }
+unpack_rgb <- function(x, filename = "") {
+  if (terra::nlyr(x) < 3) stop("cannot unpack raster with fewer than 3 layers")
+  terra::lapp(x[[1:3]], function(.x1, .x2, .x3) -10000 + ((.x1 * 256 * 256 + .x2 * 256 + .x3) * 0.1), filename = filename)
+}
+#' @name get_tiles
+#' @export
+read_tiles <- function(x, buffer, type = "mapbox.satellite", crop_to_buffer = TRUE,
+                      format = NULL, ..., zoom = NULL, debug = FALSE, max_tiles = NULL, base_url = NULL,
+                      verbose = TRUE, filename = "") {
+  tiles <- get_tiles(x = x, buffer = buffer, type = type, crop_to_buffer = crop_to_buffer, 
+                     format = format, ..., zoom = zoom, debug = debug, max_tiles = max_tiles, base_url = base_url, 
+                     verbose = verbose)
+  
+  ## bit of stuff from ceramic_tiles which we should expose properly
+  tile_index <- strsplit(str_extract(tiles$files, "[0-9]+/[0-9]+/[0-9]+"), "/")
+  tile_index <- do.call(rbind, lapply(tile_index, as.integer))
+  tile_x <- as.numeric(tile_index[,2L, drop = TRUE])
+  tile_y <- as.numeric(tile_index[,3L, drop = TRUE])
+  zooms <- as.numeric(tile_index[,1L, drop = TRUE])
+  
+  tile_index <- add_extent(tibble::tibble(tile_x = tile_x, tile_y = tile_y, zoom = zooms, fullname = tiles$files))
+  
+  mk_rast <- function(row) {
+    ex <- terra::ext(row$xmin, row$xmax, row$ymin, row$ymax)
+    suppressWarnings(r <- terra::rast(row$fullname))
+    set.ext(r, ex)
+    set.crs(r, .merc())
+    r
+  }
+  terra::merge(sprc(lapply(split(tile_index, 1:nrow(tile_index)), mk_rast)), filename = filename)
+}
+#' Download imagery tiles
 #'
 #' Obtain imagery or elevation tiles by location query. The first argument
 #' `loc` may be a spatial object (sp, raster, sf) or a 2-column matrix with a single
@@ -23,12 +70,10 @@ guess_format <- function(x) {
 #'
 #' Use `debug = TRUE` to avoid download and simply report on what would be done.
 #'
-#' `cc_elevation` does extra work to unpack the DEM tiles from the RGB format.
-#'
 #' Available types are 'elevation-tiles-prod' for AWS elevation tiles, and 'mapbox.satellite',
-#' 'mapbox.outdoors', 'mapbox.terrain-rgb', 'mapbox.streets', 'mapbox.light', 'mapbox.dark'
-#'  or any other string accepted by Mapbox services.
+#' 'mapbox.terrain-rgb'.  (The RGB terrain values are not unpacked.)
 #'
+#' Function `read_tiles()` will match what `get_tiles()` does and actually build a raster object. 
 #'
 #' @param x a longitude, latitude pair of coordinates, or a spatial object
 #' @param buffer width in metres to extend around the location, ignored if 'x' is a spatial object with extent
@@ -37,18 +82,19 @@ guess_format <- function(x) {
 #' @param format tile format to use, defaults to "jpg" for Mapbox satellite imagery and "png" otherwise
 #' @param ... arguments passed to internal function, specifically `base_url` (see Details)
 #' @param zoom desired zoom for tiles, use with caution - if `NULL` is chosen automatically
-#' @param debug optionally print out files that will be used
+#' @param debug optionally avoid actual download, but print out what would be downloaded in non-debug mode
 #' @param max_tiles maximum number of tiles - if `NULL` is set by zoom constraints
 #' @param base_url tile provider URL expert use only
 #' @param verbose report messages or suppress them
+#' @param filename purely for [read_tiles()] this is passed down to the the terra package function 'merge'
 #' @export
 #' @return A list with files downloaded in character vector, a data frame of the tile indices,
-#' the zoom level used and the extent in [raster::extent] form.
+#' the zoom level used and the extent in xmin,xmax,ymin,ymax form.
 #' @name get_tiles
-#' @seealso get_tiles_zoom get_tiles_dim get_tiles_buffer
+#' @seealso get_tiles_zoom get_tiles_dim get_tiles_buffer 
 #' @examples
 #' if (!is.null(get_api_key())) {
-#'    tile_info <- get_tiles(raster::extent(146, 147, -43, -42), type = "mapbox.outdoors", zoom = 5)
+#'    tile_info <- get_tiles(ext(146, 147, -43, -42), type = "mapbox.satellite", zoom = 5)
 #' }
 get_tiles <- function(x, buffer, type = "mapbox.satellite", crop_to_buffer = TRUE,
                       format = NULL, ..., zoom = NULL, debug = FALSE, max_tiles = NULL, base_url = NULL,
@@ -74,7 +120,9 @@ get_tiles <- function(x, buffer, type = "mapbox.satellite", crop_to_buffer = TRU
 
   my_bbox <- bbox_pair$tile_bbox
   bb_points <- bbox_pair$user_points
-
+ if (is.null(zoom) && is.null(max_tiles)) {
+      max_tiles <- 32
+    }
 
   tile_grid <- slippymath::bbox_to_tile_grid(my_bbox, max_tiles = max_tiles, zoom = zoom)
   zoom <- tile_grid$zoom
@@ -89,15 +137,16 @@ get_tiles <- function(x, buffer, type = "mapbox.satellite", crop_to_buffer = TRU
     query_string <- mk_query_string_custom(baseurl = base_url)
   }
 
-
+#print(query_string)
   files <- unlist(down_loader(tile_grid, query_string, debug = debug, verbose = verbose))
   bad <- file.info(files)$size < 35
-  if (all(bad)) {
+
+  if (!debug && all(bad)) {
     mess <-paste(files, collapse = "\n")
     stop(sprintf("no sensible tiles found, check cache?\n%s", mess))
   }
   user_ex <- NULL
-  if (crop_to_buffer) user_ex <- raster::extent(as.vector(bb_points))
+  if (crop_to_buffer) user_ex <- as.vector(bb_points)
   out <- list(files = files[!bad], tiles = tile_grid, extent = user_ex)
   if (debug) {
     out <- invisible(out)
@@ -127,17 +176,17 @@ get_tiles <- function(x, buffer, type = "mapbox.satellite", crop_to_buffer = TRU
 #' @name get-tiles-constrained
 #' @aliases get_tiles_zoom get_tiles_dim get_tiles_buffer
 #' @return A list with files downloaded in character vector, a data frame of the tile indices,
-#' the zoom level used and the extent in [raster::extent] form.
+#' the zoom level used and the extent in xmin,xmax,ymin,ymax form.
 #' @export
 #' @seealso get_tiles
 #' @examples
 #' if (!is.null(get_api_key())) {
-#'  ex <- raster::extent(146, 147, -43, -42)
-#'  tile_infoz <- get_tiles_zoom(ex, type = "mapbox.outdoors", zoom = 1)
+#'  ex <- ext(146, 147, -43, -42)
+#'  tile_infoz <- get_tiles_zoom(ex,  zoom = 1)
 #'
-#'  tile_infod <- get_tiles_dim(ex, type = "mapbox.outdoors", dim = c(256, 256))
+#'  tile_infod <- get_tiles_dim(ex,  dim = c(256, 256))
 #'
-#'  tile_infob <- get_tiles_buffer(cbind(146.5, -42.5), buffer = 5000, type = "mapbox.outdoors")
+#'  tile_infob <- get_tiles_buffer(cbind(146.5, -42.5), buffer = 5000)
 #' }
 get_tiles_zoom <- function(x, zoom = 0, ..., format = "png") {
   if ("max_tiles" %in% names(list(...))) {