+++ R/provider-mapbox.R
@@ -5,7 +5,7 @@ token_url <- function(api = "mapbox") {
 instruct_on_key_creation <- function(api = "mapbox") {
   if (api == "mapbox") {
     out <- paste(sprintf("To set your Mapbox API key obtain a key from %s\n", token_url()),
-    sprintf("1) Run this to set for the session 'Sys.setenv(MAPBOX_API_KEY=<yourkey>)'\n\nOR,\n\n2) To set peramanently store 'MAPBOX_API_KEY=<yourkey>' in ~/.Renviron\n\nSee 'help(ceramic::get_api_key)'"), sep = "\n")
+    sprintf("1) Run this to set for the session 'Sys.setenv(MAPBOX_API_KEY=<yourkey>)'\n\nOR,\n\n2) To set permanently store 'MAPBOX_API_KEY=<yourkey>' in ~/.Renviron\n\nSee 'help(ceramic::get_api_key)'"), sep = "\n")
   } else {
     message(sprintf("don't know if key is needed for %s", api))
   }
@@ -33,24 +33,29 @@ instruct_on_key_creation <- function(api = "mapbox") {
 #' sought in that order).
 #'
 #' If no key is available, `NULL` is returned, with a warning.
+#'
 #' @param api character string denoting which service ("mapbox" only)
+#' @param silent run in completely silent mode, default is to provide a warning
 #' @param ... currently ignored
+#'
 #' @return The stored API key value, see Details.
 #' @export
 #' @examples
 #' get_api_key()
-get_api_key <- function(api = "mapbox", ...) {
+get_api_key <- function(api = "mapbox", ..., silent = FALSE) {
   key <- NULL
   if (api == "mapbox") {
     ## Try mapdeck first (why not)
     key <- getOption("mapdeck")[['mapdeck']][[api]]
     key_candidates <- c("MAPBOX_API_KEY", "MAPBOX_API_TOKEN", "MAPBOX_KEY", "MAPBOX_TOKEN", "MAPBOX")
-    if(is.null(key)) {
-       key <- unlist(lapply(key_candidates, function(key) Sys.getenv(key)))[1L]
+    if(is.na(key) || is.null(key) || nchar(key) < 1) {
+       key <- unlist(lapply(key_candidates, function(label) Sys.getenv(label)))[1L]
     }
-    if (is.null(key) || nchar(key) < 1) {
+    if (is.na(key) || is.null(key) || nchar(key) < 1) {
       mess <- instruct_on_key_creation()
-      warning(sprintf("no mapbox key found\n\n%s", mess))
+      if (!silent) {
+        warning(sprintf("no mapbox key found\n\n%s", mess))
+      }
       key <- NULL
     }
   }